+++ tests/macro-tests/src/lib.rs
@@ -30,3 +30,33 @@ macro_rules! test_feature_vec {
         }
     };
 }
+
+#[macro_export]
+macro_rules! expand_to_valid_code {
+    () => {
+        let x = 1 + 2;
+    };
+}
+
+#[cfg(feature = "test-feature")]
+#[cfg_attr(feature = "test-feature", macro_export)]
+macro_rules! feature_expand_to_valid_code {
+    () => {
+        let x = 1 + 2;
+    };
+}
+
+#[macro_export]
+macro_rules! expand_to_invalid_code {
+    () => {
+        let x = 1 + "2";
+    };
+}
+
+#[cfg(feature = "test-feature")]
+#[cfg_attr(feature = "test-feature", macro_export)]
+macro_rules! feature_expand_to_invalid_code {
+    () => {
+        let x = 1 + "2";
+    };
+}