+++ tests/macro-tests/tests/expand_checking/fail/invalid.expand.out.rs
@@ -1,0 +1,5 @@
+#[macro_use]
+extern crate macro_tests;
+pub fn main() {
+    let x = 1 + "2";
+}