+++ tests/macro-tests/tests/expand_checking/fail/invalid.rs
@@ -1,0 +1,6 @@
+#[macro_use]
+extern crate macro_tests;
+
+pub fn main() {
+    expand_to_invalid_code!();
+}